etl/requirements.txt
elasticsearch==8.14.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pydantic==2.7.4
redis==5.0.7
